requirements.txt
duckdb>=1.1.0
streamlit>=1.40.0
pandas>=2.2.0
plotly>=5.24.0
python-dotenv>=1.0.1
anthropic>=0.37.0
langgraph>=0.2.0
langchain-core>=0.3.0
pytest>=8.3.0
